LayoutTests/fast/parser/fragment-with-frameset-crash.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<iframe srcdoc='<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Frameset//EN" "http://www.w3.org/TR/html4/frameset.dtd"><html></html>'></iframe>
<script>
description("Test that we do not crash when parsing a fragment that contains a frameset.");
jsTestIsAsync = true;

onload = function() {
    frames[0].document.documentElement.innerHTML = "<caption><frameset cols='50%,50%'><frame src='about:blank'/><frame src='about:blank'></frameset>";
    shouldBeEqualToString("frames[0].document.documentElement.innerHTML", '<head></head><frameset cols="50%,50%"><frame src="about:blank"><frame src="about:blank"></frameset>');
    finishJSTest();
}
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
